+++ .github/workflows/test-windows.yml
@@ -77,37 +77,37 @@ jobs:
       shell: pwsh
 
     - name: Build dependencies / libjpeg
-      run: "& .\\winbuild\\build_dep_libjpeg.cmd"
+      run: "& .\\winbuild\\build\\build_dep_libjpeg.cmd"
     - name: Build dependencies / zlib
-      run: "& .\\winbuild\\build_dep_zlib.cmd"
+      run: "& .\\winbuild\\build\\build_dep_zlib.cmd"
     - name: Build dependencies / LibTiff
-      run: "& .\\winbuild\\build_dep_libtiff.cmd"
+      run: "& .\\winbuild\\build\\build_dep_libtiff.cmd"
     - name: Build dependencies / WebP
-      run: "& .\\winbuild\\build_dep_libwebp.cmd"
+      run: "& .\\winbuild\\build\\build_dep_libwebp.cmd"
     - name: Build dependencies / FreeType
-      run: "& .\\winbuild\\build_dep_freetype.cmd"
+      run: "& .\\winbuild\\build\\build_dep_freetype.cmd"
     - name: Build dependencies / LCMS2
-      run: "& .\\winbuild\\build_dep_lcms2.cmd"
+      run: "& .\\winbuild\\build\\build_dep_lcms2.cmd"
     - name: Build dependencies / OpenJPEG
-      run: "& .\\winbuild\\build_dep_openjpeg.cmd"
+      run: "& .\\winbuild\\build\\build_dep_openjpeg.cmd"
 
     # GPL licensed; skip if building wheels
     - name: Build dependencies / libimagequant
       if: "github.event_name != 'push' || contains(matrix.python-version, 'pypy')"
-      run: "& .\\winbuild\\build_dep_libimagequant.cmd"
+      run: "& .\\winbuild\\build\\build_dep_libimagequant.cmd"
 
     # Raqm dependencies
     - name: Build dependencies / HarfBuzz
-      run: "& .\\winbuild\\build_dep_harfbuzz.cmd"
+      run: "& .\\winbuild\\build\\build_dep_harfbuzz.cmd"
     - name: Build dependencies / FriBidi
-      run: "& .\\winbuild\\build_dep_fribidi.cmd"
+      run: "& .\\winbuild\\build\\build_dep_fribidi.cmd"
     - name: Build dependencies / Raqm
-      run: "& .\\winbuild\\build_dep_libraqm.cmd"
+      run: "& .\\winbuild\\build\\build_dep_libraqm.cmd"
 
     - name: Build Pillow
       run: |
         cd $env:GITHUB_WORKSPACE
-        & winbuild\build_pillow.cmd install
+        & winbuild\build\build_pillow.cmd install
         & $env:pythonLocation\python.exe selftest.py --installed
       shell: pwsh
 
@@ -145,14 +145,14 @@ jobs:
       with:
         file: ./coverage.xml
         flags: GHA_Windows
-        name: ${{ runner.os }} Python ${{ matrix.python-version }}
+        name: ${{ runner.os }} Python ${{ matrix.python-version }} ${{ matrix.architecture }}
 
     - name: Build wheel
       id: wheel
       if: "github.event_name == 'push' && !contains(matrix.python-version, 'pypy')"
       run: |
         for /f "tokens=3 delims=/" %%a in ("${{ github.ref }}") do echo ::set-output name=dist::dist-%%a
-        %GITHUB_WORKSPACE%\\winbuild\\build_pillow.cmd bdist_wheel"
+        %GITHUB_WORKSPACE%\\winbuild\\build\\build_pillow.cmd bdist_wheel"
       shell: cmd
 
     - uses: actions/upload-artifact@v1
@@ -160,9 +160,3 @@ jobs:
       with:
         name: ${{ steps.wheel.outputs.dist }}
         path: dist
-
-#    - uses: actions/upload-artifact@v1
-#      if: matrix.architecture == 'x86'
-#      with:
-#        name: lib
-#        path: "winbuild\\build\\3.x\\x86\\lib"
